fix(anyharness): keep producer/mod.rs under the max-lines gate
CI repo-shape failed: the dev-refresh additions pushed producer/mod.rs to
644 lines (max 600). Relocate all dev-refresh glue (spawn helper, dev-only
ProducerInner accessors, test module registration) into the cfg-gated
dev_refresh.rs, and move mark_bridge_lost next to its only caller in
bridge_runtime.rs. No behavior change; 198 crate tests stay green.
