+++ anyharness/crates/proliferate-diagnostics-client/src/producer/mod.rs
@@ -5,7 +5,7 @@ use std::{
 };
 use tokio::time::Instant as TokioInstant;
 
-use proliferate_diagnostics_protocol::v1::{limits::MAX_SAFE_INTEGER, types::ProducerRecordV1};
+use proliferate_diagnostics_protocol::v1::types::ProducerRecordV1;
 
 use crate::{
     bridge::activation::{BundledDesktopDiagnosticsBootstrap, InitialCollectorState},
@@ -27,9 +27,6 @@ pub(crate) mod status;
 pub(crate) mod transport;
 mod worker;
 
-#[cfg(all(test, unix, debug_assertions))]
-#[path = "tests_dev_refresh.rs"]
-mod tests_dev_refresh;
 #[cfg(all(test, unix))]
 #[path = "tests_delivery_end.rs"]
 mod tests_delivery_end;
@@ -192,8 +189,6 @@ pub(crate) fn install(
             bootstrap.bridge.zip(bootstrap.shutdown),
         ),
         // No inherited descriptors: no bridge thread, no fallback authority.
-        // The file-backed refresh task below stands in for the bridge's
-        // generation updates when the host published the snippet's path.
         #[cfg(debug_assertions)]
         BundledDesktopDiagnosticsBootstrap::DevEnv(dev) => {
             dev_env_path = dev.env_path;
@@ -215,11 +210,7 @@ pub(crate) fn install(
             Some(bootstrap.classification),
         ),
         #[cfg(debug_assertions)]
-        BundledDesktopDiagnosticsBootstrap::DevEnv(dev) => {
-            // The file-backed refresh is unix-only, like the fd bridge.
-            let _ = &dev.env_path;
-            (dev.initial_state, None, None)
-        }
+        BundledDesktopDiagnosticsBootstrap::DevEnv(dev) => (dev.initial_state, None, None),
     };
     if degraded.is_some() {
         eprintln!("[desktop-diagnostics] bundled bootstrap degraded");
@@ -292,12 +283,7 @@ pub(crate) fn install(
         .transpose()
         .map_err(|_| InstallError::WorkerUnavailable)?;
     #[cfg(all(unix, debug_assertions))]
-    if let Some(path) = dev_env_path {
-        runtime.spawn(dev_refresh::dev_generation_refresh(
-            Arc::clone(&inner),
-            path,
-        ));
-    }
+    dev_refresh::spawn_if_configured(&runtime, &inner, dev_env_path);
     let join = runtime.spawn(worker::run(Arc::clone(&inner)));
     let handle = DiagnosticsProducerHandle {
         inner: Arc::clone(&inner),
@@ -444,29 +430,6 @@ impl ProducerInner {
         self.notify.notify_one();
     }
 
-    /// Boot id of the current generation, ready or cooling down. `None` while
-    /// unavailable so a re-published snippet for the same (dead) collector can
-    /// still be retried by the dev refresh loop.
-    #[cfg(all(unix, debug_assertions))]
-    pub(crate) fn current_collector_boot_id(&self) -> Option<String> {
-        let state = self.state.lock().unwrap_or_else(|error| error.into_inner());
-        match &state.collector {
-            CollectorAvailability::Ready(generation)
-            | CollectorAvailability::Cooldown { generation, .. } => {
-                Some(generation.collector_boot_id.clone())
-            }
-            CollectorAvailability::Unavailable { .. } => None,
-        }
-    }
-
-    #[cfg(all(unix, debug_assertions))]
-    pub(crate) fn is_terminal(&self) -> bool {
-        self.state
-            .lock()
-            .unwrap_or_else(|error| error.into_inner())
-            .terminal
-    }
-
     #[cfg(unix)]
     pub(crate) fn mark_generation_unavailable(&self, generation: u64) {
         let mut state = self.state.lock().unwrap_or_else(|error| error.into_inner());
@@ -501,27 +464,6 @@ impl ProducerInner {
         // The parent signal transfers clock ownership immediately; dispatch
         // pauses until the request supplies the parent's remaining window.
         state.terminal_deadline = None;
-        drop(state);
-        self.notify.notify_one();
-    }
-
-    /// Permanent bridge loss: no newer generation can arrive, so the sentinel
-    /// latches unavailable. Queued records retain routing until worker drain.
-    #[cfg(unix)]
-    pub(crate) fn mark_bridge_lost(&self) {
-        let mut state = self.state.lock().unwrap_or_else(|error| error.into_inner());
-        if matches!(
-            &state.collector,
-            CollectorAvailability::Unavailable { generation } if *generation == MAX_SAFE_INTEGER
-        ) {
-            return;
-        }
-        state.collector = CollectorAvailability::Unavailable {
-            generation: MAX_SAFE_INTEGER,
-        };
-        if !state.in_flight.is_empty() {
-            state.delivery_fence_eligible = false;
-        }
         drop(state);
         self.notify.notify_one();
     }
